requirements.txt
allennlp==0.8.4
pytorch-pretrained-bert==0.6.2
docopt==0.6.2
